uploaded data including CSV on ammonia emissions
I don't know if it actually connects at all with CO2, but according to RAND.org, "Ammonia is one of the main sources of nitrogen pollution, alongside nitrogen oxides. A major effect of ammonia pollution on biodiversity is the impact of nitrogen accumulation on plant species diversity and composition within affected habitats. "